requirements.txt
torch==2.2.1
torchvision==0.17.1
transformers>=4.37.2
accelerate>=0.27.1
bitsandbytes>=0.42.0
huggingface-hub>=0.20.3
pillow>=10.2.0
protobuf>=4.25.2
safetensors>=0.4.2
sentencepiece>=0.1.99
tokenizers>=0.15.2
tqdm>=4.66.2
